golfconnectx/requirements.txt
click==6.6
decorator==4.0.10
Django==1.8
django-cors-headers==1.3.1
django-cors-middleware==1.3.1
django-haystack==2.6.0
django-pipeline==1.6.9
django-pipeline-browserify==0.5.0
djangorestframework==3.5.3
drf-haystack==1.6.1
easy-thumbnails==2.3
elasticsearch==5.1.0
geocoder==1.19.0
googlemaps==2.4.5
httpie==0.9.6
Pillow==3.4.2
psycopg2==2.6.2
pygeocoder==1.2.5
Pygments==2.1.3
python-dateutil==2.6.0
python-thumbnails==0.5.1
ratelim==0.1.6
requests==2.6.0
six==1.9.0
urllib3==1.20
